+++ app/services/admin/owners/drop_contacts.rb
@@ -1,0 +1,6 @@
+class Admin::Owners::DropContacts < Admin::Owners
+  def do
+    @owner.contacts.destroy_all
+    [true, 'All contacts was destroyed']
+  end
+end