+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+### v0.6.6
+- :bulb: Added feature to clean all contacts statuses
+- :bulb: Added hidden feature to mark contacts and friends randomly
+- :arrow_up: Refactored danger controller actions
+- :arrow_up: Added specs for add/ignore contact services
+
 ### v0.6.5
 - :bulb: Implemented mobile design for web client page
 - :arrow_up: Changed assets precompiler from SASS to SCSS